READMe.md
Is Tech all that you are looking for? Are you tired of endlessly searching for Tech friends who can offer real-time help when needed ? Then i've got your back buddy Sam  over here ,  A Technologist , programmer and a cyber security enthusiast . Feel free to collaborate with me in your projects you can pull , review and collaborate all you can with the contents that i will be posting here.
I will be writing and posting so many scripts that in a way will be able to solve some of your problems. FEEL FREE TO COPY , MODIFY AND REDISTRIBUTE THE CODE (But don't FORGET to ACKNOWLEDGE my work. *******************************************************